requirements.txt
sanic
pymodm
passlib
pyjwt
marshmallow
celery
